Move Stop out of uswds directory
Jira ticket: CAMS-510
